CHG: Added ifdefs to make pwRecon work on MacOs.
NEW: Added a self compiled version of hashcat because there is no build version for MacOs available right now.

INFO: pwRecon is not working properly right now because of the new Hashcat version and the missing support for the local passwords.

INFO: The build process is creating a .app but the tools and the translations are missing. Still need to add those manually.